[hotfix][network][tests] split PipelinedSubpartitionTest for better initialization
- add PipelinedSubpartitionWithReadViewTest which always creates a subpartition,
an availability listener, and a read view before each test and cleans up after
each test
- remove mockito use from testBasicPipelinedProduceConsumeLogic()
